#1202 Fix more PR suggestions. Moved Kafka-independent plugin out of Kafka package.
