+++ spark-jobs/src/main/resources/application.conf.template
@@ -59,8 +59,8 @@ timezone="UTC"
 
 # 1. Push Control Info measurements to a Kafka topic
 # (uncomment to enable)
-#standardization.plugin.control.metrics.1=za.co.absa.enceladus.plugins.builtin.controlinfo.kafka.KafkaInfoPlugin
-#conformance.plugin.control.metrics.1=za.co.absa.enceladus.plugins.builtin.controlinfo.kafka.KafkaInfoPlugin
+#standardization.plugin.control.metrics.1=za.co.absa.enceladus.plugins.builtin.controlinfo.mq.kafka.KafkaInfoPlugin
+#conformance.plugin.control.metrics.1=za.co.absa.enceladus.plugins.builtin.controlinfo.mq.kafka.KafkaInfoPlugin
 #kafka.schema.registry.url:"http://127.0.0.1:8081"
 #kafka.bootstrap.servers="127.0.0.1:9092"
 #kafka.info.metrics.client.id="controlInfo"
